finance-genie/multi-agent-demo/pyproject.toml
[project]
name = "finance-genie-multi-agent-demo"
version = "0.1.0"
description = "Finance Genie Neo4j GDS fraud specialist for Supervisor Agent demos."
requires-python = ">=3.12"
dependencies = [
    "databricks-agents>=1.1.0",
    "databricks-job-runner>=0.4.3",
    "databricks-langchain",
    "databricks-mcp",
    "databricks-sdk>=0.40",
    "langchain-core",
    "langgraph",
    "langgraph-prebuilt",
    "mcp",
    "mlflow>=3.1.3",
    "nest-asyncio",
    "pydantic>=2",
    "python-dotenv",
    "requests",
]

[tool.uv]
package = false
